Install conda build and deployment tools directly
Instead of using Obvious-CI to install these build tools, just install
them directly. At this point, it appears we are just using an Obvious-CI
script as a wrapper around a `conda install` command. We don't appear to
be using anything else from Obvious-CI in the Docker image. So simply
drop Obvious-CI and do the install directly.