Ensure `/home/conda` is owned by `conda`
Normally `/home/conda` is owned by `conda` if the directory is created
through `useradd`. However it might not be owned by `conda` if a volume
is mounted by Docker to some path that requires `/home/conda` already
exist. In this case, we ensure that `/home/conda` is owned by the
`conda` user and `conda` group by a call to `chown` after creating the
`conda` user.