Copy contents from `/etc/skel`
Normally `useradd` will copy contents from `/etc/skel` to the user's
directory when creating it. However this won't happen if the user's
directory already exists, which can happen if a directory is mounted by
Docker into the user's directory. To fix this issue, add a line to the
`entrypoint` script to copy over the contents of `/etc/skel` to the
user's home directory manually (fixing ownership on the way).