Use force when removing `miniconda.sh`
As this is a file downloaded by `root`, the system will ask us to
confirm our decision to remove the file. Since we are running in batch,
it simply assumes the answer is no and proceeds. As a result, a copy of
`miniconda.sh` (weighing in at about 34MB) is getting packaged in our
Docker image increasing the image size.

To fix this, start using `rm -f` to forcibly remove `miniconda.sh`. This
way when Docker goes to commit this layer, `miniconda.sh` won't be
present and so will not be included. Should help improve image size and
cutdown on clutter.