Remove pre-included system compilers
As `nvcc` may try to use these ancient compilers before using the
Anaconda compilers, go ahead and manually remove them. Though make
sure that `nvcc` is not uninstalled in the process as it requires
`gcc` and `gcc-c++`. Do this by calling `rpm` and asking it to remove
the package without prompting and without any dependent packages.