Enable docker's experimental features
In order to get access to the `--squash` flag in `docker build`, we need
to enable experimental features. This change supplies the `daemon.json`
configuration file we want docker to use and moves it in place. Docker
needs to be restarted for this feature to take effect, which we handle
after.